refactor: organize knowhow and vote queries
