+++ src/app/(main)/boards/knowhow/_components/KnowhowEditor.tsx
@@ -5,8 +5,8 @@ import "react-quill/dist/quill.snow.css";
 import dynamic from "next/dynamic";
 import Button from "@/components/Button/Button";
 import ReactQuill, { ReactQuillProps } from "react-quill";
-import useKnowhowImageMutation from "@/stores/queries/useKnowhowImageMutation";
-import useKnowhowMutation from "@/stores/queries/useKnowhowMutation";
+import useKnowhowImageMutation from "@/stores/queries/knowhow/post/useKnowhowImageMutation";
+import useKnowhowMutation from "@/stores/queries/knowhow/post/useKnowhowMutation";
 import { TKnowhow } from "@/types/knowhow.type";
 import { useUserContext } from "@/provider/contexts/UserContext";
 import { useModal } from "@/provider/contexts/ModalContext";
