+++ src/stores/queries/knowhow/post/useKnowhowImageMutation.ts
