+++ src/stores/queries/vote/post/useVoteMutation.ts
